Added Check for comms and changed us for servo

+++ Source/SolarScavenger/include/comms.h
@@ -51,7 +51,7 @@ private:
 
     uint8_t receiverMac[6];
 
-    uint32_t lastMessageMicros;
+    uint64_t lastMessageMicros;
 
 };
